internal/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
anyio==3.7.1
attrs==25.4.0
blessed==1.22.0
certifi==2025.10.5
charset-normalizer==3.4.4
codefind==0.1.7
cogwatch==3.3.1
contourpy==1.3.3
cycler==0.12.1
discord.py==2.6.4
dnspython==2.8.0
dotenv==0.9.9
fonttools==4.60.1
frozenlist==1.8.0
idna==3.11
jurigged==0.6.1
kiwisolver==1.4.9
matplotlib==3.10.7
multidict==6.7.0
numpy==2.3.4
osrparse==7.0.1
osu.py==4.1.0
ovld==0.5.13
packaging==25.0
pillow==12.0.0
propcache==0.4.1
pymongo==4.15.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
watchdog==6.0.0
watchfiles==0.15.0
wcwidth==0.2.14
websockets==13.1
yarl==1.22.0
